refactoring: esm.mjs solution
